support for replacing template comments with top-level elements (such as content from a separate document) like paragraphs or tables
